fix various ambiguity (#65)
* add aqua test

* fix ambiguity on getindex

* Revert "add aqua test"

This reverts commit 2d2f2a0f95e6a555cb5d1e46529aa97a3cc38d83.

+++ src/SSparseMatrixCSC.jl
@@ -77,7 +77,7 @@ function size(spm::SSparseMatrixCSC{Tv,Ti,NNZ,NP}, i::Integer) where {Tv,Ti,NNZ,
 end
 size(spm::SSparseMatrixCSC{Tv,Ti,NNZ,NP}) where {Tv,Ti,NNZ,NP} = (spm.m, NP - 1)
 
-function getindex(ssp::SSparseMatrixCSC{Tv}, i, j) where {Tv}
+function getindex(ssp::SSparseMatrixCSC{Tv}, i::Integer, j::Integer) where {Tv}
     S = ssp.colptr[j]
     E = ssp.colptr[j+1] - 1
     for ii = S:E
